Wallet Application
You will be creating a simple wallet application which allows you to add
and remove amounts to
and from your wallet. Once an amount is entered
into the wallet, it should be listed along with the
date it was added.
Once an amount is removed from the wallet it should be listed along with
the
date it was removed. There should then be a grand total displaying
how much money is currently
in the wallet. There will be a responsive
header with three menu items: Home, Reset, View
Source.